+++ RTCP/Cobalt/GPUProc/src/cuda/Kernels/IncoherentStokesKernel.h
@@ -43,10 +43,10 @@ namespace LOFAR
       struct Parameters : Kernel::Parameters
       {
         Parameters(const Parset& ps);
-        size_t nrChannelsPerSubband;
-        size_t nrSamplesPerChannel;
-        size_t nrStokes;
-        size_t timeIntegrationFactor;
+        unsigned nrChannelsPerSubband;
+        unsigned nrSamplesPerChannel;
+        unsigned nrStokes;
+        unsigned timeIntegrationFactor;
       };
 
       enum BufferType
